Rename di_injection.dart to dependency_injection.dart for consistency; update imports in main.dart and news_list_view.dart
