+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:news_app/core/utils/di_injection.dart';
+import 'package:news_app/core/utils/dependency_injection.dart';
 import 'package:news_app/core/utils/simple_observer.dart';
 import 'package:news_app/features/home/presentation/views/home_view.dart';
